s/processd: restructure the new code
