+++ services/processd/include/pe.h
@@ -73,11 +73,7 @@ typedef struct PeExecutable {
 /*******************************************************************************
  * Support Methods 
  *******************************************************************************/
-__EXTERN uintptr_t  PECurrentPageSize(void);
 __EXTERN uintptr_t  PeImplGetBaseAddress(void);
-__EXTERN clock_t    PeImplGetTimestampMs(void);
-__EXTERN oserr_t PeImplResolveFilePath(uuid_t, mstring_t*, mstring_t*, mstring_t**);
-__EXTERN oserr_t __LoadFile(mstring_t *fullPath, void **bufferOut, size_t *lengthOut);
 __EXTERN oserr_t PeImplCreateImageSpace(uuid_t* handleOut);
 
 /*******************************************************************************
