+++ services/processd/pe/mapper.c
@@ -24,301 +24,8 @@
 #include <os/memory.h>
 #include <string.h>
 #include <stdlib.h>
-#include "pe.h"
-#include "mapper.h"
+#include <module.h>
 #include "private.h"
-
-// Reuse the ordinal_cmp for name as well, the hashtable only calls cmp
-// once hashes match, and if hashes match on function names, we just do
-// an ordinal compare.
-static uint64_t __expfn_ordinal_hash(const void* element);
-static int      __expfn_ordinal_cmp(const void* element1, const void* element2);
-static uint64_t __expfn_name_hash(const void* element);
-
-struct Module*
-ModuleNew(
-        _In_ void*  moduleBuffer,
-        _In_ size_t bufferSize)
-{
-    struct Module* module;
-    int            status;
-
-    module = malloc(sizeof(struct Module));
-    if (module == NULL) {
-        return NULL;
-    }
-    memset(module, 0, sizeof(struct Module));
-
-    module->ImageBuffer = moduleBuffer;
-    module->ImageBufferSize = bufferSize;
-    usched_mtx_init(&module->Mutex, USCHED_MUTEX_PLAIN);
-    status = hashtable_construct(
-            &module->ExportedOrdinals, 0, sizeof(struct ExportedFunction),
-            __expfn_ordinal_hash, __expfn_ordinal_cmp);
-    if (status) {
-        free(module);
-        return NULL;
-    }
-    status = hashtable_construct(
-            &module->ExportedNames, 0, sizeof(struct ExportedFunction),
-            __expfn_name_hash, __expfn_ordinal_cmp);
-    if (status) {
-        hashtable_destroy(&module->ExportedOrdinals);
-        free(module);
-        return NULL;
-    }
-    return module;
-}
-
-void
-ModuleDelete(
-        _In_ struct Module* module)
-{
-    if (module == NULL) {
-        return;
-    }
-
-    // No further cleanup is needed for exported functions, as no allocations
-    // are made for the structs themselves.
-    hashtable_destroy(&module->ExportedOrdinals);
-    hashtable_destroy(&module->ExportedNames);
-    free(module->Sections);
-    free(module->ImageBuffer);
-    free(module);
-}
-
-static void
-__CopyDirectories(
-        _In_ struct Module*     module,
-        _In_ PeDataDirectory_t* directories)
-{
-    memcpy(
-            &module->DataDirectories[0],
-            directories,
-            sizeof(PeDataDirectory_t) * PE_NUM_DIRECTORIES
-    );
-}
-
-static void
-__ParsePE32Headers(
-        _In_  struct Module*      module,
-        _In_  PeOptionalHeader_t* optionalHeader,
-        _Out_ void**              sectionHeadersOut)
-{
-    PeOptionalHeader32_t* header = (PeOptionalHeader32_t*)optionalHeader;
-
-    module->ImageBase    = header->BaseAddress;
-    module->MetaDataSize = header->SizeOfHeaders;
-    __CopyDirectories(module, (PeDataDirectory_t*)&header->Directories[0]);
-
-    *sectionHeadersOut = (void*)((uint8_t*)header + sizeof(PeOptionalHeader32_t));
-}
-
-static void
-__ParsePE64Headers(
-        _In_  struct Module*      module,
-        _In_  PeOptionalHeader_t* optionalHeader,
-        _Out_ void**              sectionHeadersOut)
-{
-    PeOptionalHeader64_t* header = (PeOptionalHeader64_t*)optionalHeader;
-
-    module->ImageBase    = (uintptr_t)header->BaseAddress;
-    module->MetaDataSize = header->SizeOfHeaders;
-    __CopyDirectories(module, (PeDataDirectory_t*)&header->Directories[0]);
-
-     *sectionHeadersOut = (void*)((uint8_t*)header + sizeof(PeOptionalHeader64_t));
-}
-
-static oserr_t
-__ParseModuleHeaders(
-        _In_  struct Module* module,
-        _Out_ void**         sectionHeadersOut,
-        _Out_ int*           sectionCountOut)
-{
-    MzHeader_t*         dosHeader;
-    PeHeader_t*         peHeader;
-    PeOptionalHeader_t* optionalHeader;
-
-    // Avoid doing any further checks on DOS/PE headers as they have been validated
-    // earlier. We do however match against current machine and architecture
-    dosHeader      = (MzHeader_t*)module->ImageBuffer;
-    peHeader       = (PeHeader_t*)((uint8_t*)module->ImageBuffer + dosHeader->PeHeaderAddress);
-    if (peHeader->Machine != PE_CURRENT_MACHINE) {
-        ERROR("__ParseModuleHeaders The image as built for machine type 0x%x, "
-              "which is not the current machine type.", peHeader->Machine);
-        return OS_EUNKNOWN;
-    }
-    *sectionCountOut = peHeader->NumSections;
-
-    optionalHeader = (PeOptionalHeader_t*)((uint8_t*)peHeader + sizeof(PeHeader_t));
-    if (optionalHeader->Architecture != PE_CURRENT_ARCH) {
-        ERROR("__ParseModuleHeaders The image was built for architecture 0x%x, "
-              "and was not supported by the current architecture.", optionalHeader->Architecture);
-        return OS_EUNKNOWN;
-    }
-
-    if (optionalHeader->Architecture == PE_ARCHITECTURE_32) {
-        __ParsePE32Headers(module, optionalHeader, sectionHeadersOut);
-    } else if (optionalHeader->Architecture == PE_ARCHITECTURE_64) {
-        __ParsePE64Headers(module, optionalHeader, sectionHeadersOut);
-    } else {
-        ERROR("__ParseModuleHeaders Unsupported architecture %x", optionalHeader->Architecture);
-        return OS_EUNKNOWN;
-    }
-    return OS_EOK;
-}
-
-static unsigned int
-__GetSectionPageFlags(
-        _In_ PeSectionHeader_t* section)
-{
-    unsigned int flags = MEMORY_READ;
-    if (section->Flags & PE_SECTION_EXECUTE) {
-        flags |= MEMORY_EXECUTABLE;
-    }
-    if (section->Flags & PE_SECTION_WRITE) {
-        flags |= MEMORY_WRITE;
-    }
-    return flags;
-}
-
-static oserr_t
-__ParseModuleSections(
-        _In_ struct Module* module,
-        _In_ void*          sectionHeadersData,
-        _In_ int            sectionCount)
-{
-    PeSectionHeader_t* section = sectionHeadersData;
-
-    module->Sections = malloc(sizeof(struct Section) * sectionCount);
-    if (module->Sections == NULL) {
-        return OS_EOOM;
-    }
-    memset(module->Sections, 0, sizeof(struct Section) * sectionCount);
-
-    for (int i = 0; i < sectionCount; i++, section++) {
-        uint8_t* sectionFileData = ((uint8_t*)module->ImageBuffer + section->RawAddress);
-
-        // Calculate page flags for the section
-        memcpy(&module->Sections[i].Name[0], &section->Name[0], PE_SECTION_NAME_LENGTH);
-        module->Sections[i].FileData     = (const void*)sectionFileData;
-        module->Sections[i].MapFlags     = __GetSectionPageFlags(section);
-        module->Sections[i].FileLength   = section->RawSize;
-        module->Sections[i].MappedLength = section->VirtualSize;
-        module->Sections[i].RVA          = section->VirtualAddress;
-
-        // Is it a zero section?
-        if (section->RawSize == 0 || (section->Flags & PE_SECTION_BSS)) {
-            module->Sections[i].Zero = true;
-        }
-        module->SectionCount++;
-    }
-    return OS_EOK;
-}
-
-static void*
-__ConvertRVAToDataPointer(
-        _In_ struct Module* module,
-        _In_ uint32_t       rva)
-{
-    for (int i = 0; i < module->SectionCount; i++) {
-        if (rva >= module->Sections[i].RVA && rva < (module->Sections[i].RVA + module->Sections[i].MappedLength)) {
-            return (void*)((uint8_t*)module->Sections[i].FileData + (rva - module->Sections[i].RVA));
-        }
-    }
-    return NULL;
-}
-
-static oserr_t
-__ParseModuleExportedFunctions(
-        _In_ struct Module* module)
-{
-    PeDataDirectory_t*   directory;
-    PeExportDirectory_t* exportDirectory;
-    uint32_t*            nameTable;
-    uint16_t*            ordinalTable;
-    uint32_t*            addressTable;
-
-    directory = &module->DataDirectories[PE_SECTION_EXPORT];
-    if (directory->AddressRVA == 0 || directory->Size == 0) {
-        // No exports
-        return OS_EOK;
-    }
-
-    // Get the file data which the RVA points to
-    exportDirectory = __ConvertRVAToDataPointer(
-            module,
-            module->DataDirectories[PE_SECTION_EXPORT].AddressRVA);
-    if (exportDirectory == NULL) {
-        ERROR("__ParseModuleExportedFunctions export directory was invalid");
-        return OS_EUNKNOWN;
-    }
-
-    // Ordinals and names can exceed the number of exported function addresses. This is because
-    // modules can export symbols from other modules, which may then exceed the number of actual
-    // functions exported local to the module.
-    nameTable = __ConvertRVAToDataPointer(module, exportDirectory->AddressOfNames);
-    ordinalTable = __ConvertRVAToDataPointer(module, exportDirectory->AddressOfOrdinals);
-    addressTable = __ConvertRVAToDataPointer(module, exportDirectory->AddressOfFunctions);
-    for (int i = 0; i < exportDirectory->NumberOfNames; i++) {
-        const char* name        = __ConvertRVAToDataPointer(module, nameTable[i]);
-        uint32_t    ordinal     = (uint32_t)ordinalTable[i] - exportDirectory->OrdinalBase;
-        uint32_t    fnRVA       = addressTable[ordinal];
-        const char* forwardName = NULL;
-
-        // The function RVA (fnRVA) is a field that uses one of two formats in the following table.
-        // If the address specified is not within the export section (as defined by the address and
-        // length that are indicated in the optional header), the field is an export RVA, which
-        // is an actual address in code or data. Otherwise, the field is a forwarder RVA, which
-        // names a symbol in another DLL.
-        if (fnRVA >= directory->AddressRVA && fnRVA < (directory->AddressRVA + directory->Size)) {
-            // The field is a forwarder RVA, which names a symbol in another DLL.
-            forwardName = __ConvertRVAToDataPointer(module, fnRVA);
-            fnRVA = 0; // fnRVA is invalid
-        }
-
-        hashtable_set(&module->ExportedOrdinals, &(struct ExportedFunction) {
-                .Name = name,
-                .ForwardName = forwardName,
-                .Ordinal = (int)ordinalTable[i],
-                .RVA = fnRVA
-        });
-        hashtable_set(&module->ExportedNames, &(struct ExportedFunction) {
-                .Name = name,
-                .ForwardName = forwardName,
-                .Ordinal = (int)ordinalTable[i],
-                .RVA = fnRVA
-        });
-    }
-    return OS_EOK;
-}
-
-static oserr_t
-__ParseModule(
-        _In_ struct Module* module)
-{
-    void*   sectionHeaders;
-    int     sectionCount;
-    oserr_t oserr;
-
-    oserr = __ParseModuleHeaders(module, &sectionHeaders, &sectionCount);
-    if (oserr != OS_EOK) {
-        return oserr;
-    }
-
-    oserr = __ParseModuleSections(module, sectionHeaders, sectionCount);
-    if (oserr != OS_EOK) {
-        return oserr;
-    }
-
-    oserr = __ParseModuleExportedFunctions(module);
-    if (oserr != OS_EOK) {
-        return oserr;
-    }
-
-    module->Parsed = true;
-    return OS_EOK;
-}
 
 static oserr_t
 __CreateMapping(
@@ -345,14 +52,14 @@ __DestroyMapping(
 
 static uintptr_t
 __AllocateLoadSpace(
-        _InOut_ uintptr_t* loadAddress,
-        _In_    size_t     size)
+        _In_    struct Module* module,
+        _InOut_ uintptr_t*     loadAddress,
+        _In_    size_t         size)
 {
-    // TODO: this should use SectionAlignment not page size
-    size_t    pageSize  = PECurrentPageSize();
-    size_t    pageCount = DIVUP(size, pageSize);
+    size_t    alignment = module->SectionAlignment;
+    size_t    count     = DIVUP(size, alignment);
     uintptr_t address   = *loadAddress;
-    *loadAddress += pageCount * pageSize;
+    *loadAddress += count * alignment;
     return address;
 }
 
@@ -366,7 +73,7 @@ __MapMetaData(
     oserr_t               oserr;
     TRACE("__MapMetaData()");
 
-    metaData.MappedAddress = __AllocateLoadSpace(loadAddress, module->MetaDataSize);
+    metaData.MappedAddress = __AllocateLoadSpace(module, loadAddress, module->MetaDataSize);
     metaData.Length = module->MetaDataSize;
     metaData.Flags = MEMORY_READ | MEMORY_WRITE;
     oserr = __CreateMapping(memorySpace, &metaData);
@@ -382,6 +89,7 @@ __MapMetaData(
 
 static oserr_t
 __MapSection(
+        _In_    struct Module*         module,
         _In_    struct Section*        section,
         _In_    uuid_t                 memorySpace,
         _InOut_ uintptr_t*             loadAddress,
@@ -397,7 +105,7 @@ __MapSection(
     mapping->RVA    = section->RVA;
     mapping->Length = sectionLength;
     mapping->Flags  = section->MapFlags;
-    mapping->MappedAddress = __AllocateLoadSpace(loadAddress, sectionLength);
+    mapping->MappedAddress = __AllocateLoadSpace(module, loadAddress, sectionLength);
     oserr = __CreateMapping(memorySpace, mapping);
     if (oserr != OS_EOK) {
         return oserr;
@@ -435,7 +143,13 @@ __MapSections(
     memset(mappings, 0, sizeof(struct SectionMapping) * module->SectionCount);
 
     for (int i = 0; i < module->SectionCount; i++) {
-        oserr_t oserr = __MapSection(&module->Sections[i], memorySpace, loadAddress, &mappings[i]);
+        oserr_t oserr = __MapSection(
+                module,
+                &module->Sections[i],
+                memorySpace,
+                loadAddress,
+                &mappings[i]
+        );
         if (oserr != OS_EOK) {
             free(mappings);
             return oserr;
@@ -458,7 +172,7 @@ __MapModule(
     // double init of the module in a multithreaded scenario.
     usched_mtx_lock(&module->Mutex);
     if (!module->Parsed) {
-        oserr = __ParseModule(module);
+        oserr = PEParseModule(module);
         if (oserr != OS_EOK) {
             usched_mtx_unlock(&module->Mutex);
             return oserr;
@@ -541,6 +255,19 @@ __ProcessBaseRelocationTable(
     return OS_EOK;
 }
 
+static void*
+__ConvertRVAToDataPointer(
+        _In_ struct Module* module,
+        _In_ uint32_t       rva)
+{
+    for (int i = 0; i < module->SectionCount; i++) {
+        if (rva >= module->Sections[i].RVA && rva < (module->Sections[i].RVA + module->Sections[i].MappedLength)) {
+            return (void*)((uint8_t*)module->Sections[i].FileData + (rva - module->Sections[i].RVA));
+        }
+    }
+    return NULL;
+}
+
 static oserr_t
 __ProcessBaseRelocations(
         _In_ struct Module*         module,
@@ -593,15 +320,23 @@ __ProcessModule(
 
 static struct ModuleMapping*
 __MapperModuleNew(
-        _In_ struct Module* module)
+        _In_ struct Module*         module,
+        _In_ struct SectionMapping* mappings,
+        _In_ uintptr_t              mappingBase)
 {
-    struct ModuleMapping* mapperModule;
+    struct ModuleMapping* moduleMapping;
 
-    mapperModule = malloc(sizeof(struct ModuleMapping));
-    if (mapperModule == NULL) {
+    moduleMapping = malloc(sizeof(struct ModuleMapping));
+    if (moduleMapping == NULL) {
         return NULL;
     }
-    return mapperModule;
+
+    moduleMapping->MappingBase = mappingBase;
+    moduleMapping->Mappings    = mappings;
+    moduleMapping->MappingCount = module->SectionCount;
+    moduleMapping->ModuleArchitecture = module->Architecture;
+    moduleMapping->DataDirectories = &module->DataDirectories[0];
+    return moduleMapping;
 }
 
 oserr_t
@@ -614,6 +349,7 @@ MapperLoadModule(
     struct Module*         module;
     uint32_t               moduleHash;
     oserr_t                oserr;
+    uintptr_t              baseAddress;
     uintptr_t              imageDelta;
     TRACE("MapperLoadModule(path=%ms)", path);
 
@@ -627,7 +363,8 @@ MapperLoadModule(
     }
 
     // Calculate the image delta for processings
-    imageDelta = (loadContext->LoadAddress - module->ImageBase);
+    baseAddress = loadContext->LoadAddress;
+    imageDelta  = (loadContext->LoadAddress - module->ImageBase);
 
     // Map the module into the memory space provided
     oserr = __MapModule(
@@ -648,7 +385,7 @@ MapperLoadModule(
 
     // Store the module hash so the loader can reduce the reference
     // count again later
-    *moduleMappingOut = __MapperModuleNew(module);
+    *moduleMappingOut = __MapperModuleNew(module, mappings, baseAddress);
     return OS_EOK;
 }
 
@@ -657,36 +394,4 @@ MapperUnloadModule(
         _In_ struct ModuleMapping* moduleMapping)
 {
     return OS_EOK;
-}
-
-static uint64_t __expfn_ordinal_hash(const void* element)
-{
-    const struct ExportedFunction* function = element;
-    return (uint64_t)function->Ordinal;
-}
-
-static int __expfn_ordinal_cmp(const void* element1, const void* element2)
-{
-    const struct ExportedFunction* function1 = element1;
-    const struct ExportedFunction* function2 = element2;
-    return function1->Ordinal == function2->Ordinal ? 0 : -1;
-}
-
-uint32_t __hash(const char* string)
-{
-    uint32_t hash = 5381;
-    size_t   i    = 0;
-    if (string == NULL) {
-        return 0;
-    }
-    while (string[i]) {
-        hash = ((hash << 5) + hash) + string[i]; /* hash * 33 + c */
-    }
-    return hash;
-}
-
-static uint64_t __expfn_name_hash(const void* element)
-{
-    const struct ExportedFunction* function = element;
-    return __hash(function->Name);
 }
